:bug: fix: useNavigate 수정
