+++ client/src/pages/Boast.js
@@ -1,8 +1,6 @@
 import styled from 'styled-components';
 import Card from '../components/Boast/Card';
 import TitleHeader from '../components/Shared/TitleHeader';
-import { useNavigate } from 'react-router';
-import { Link } from 'react-router-dom';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 
@@ -11,11 +9,11 @@ function Boast() {
   useEffect(() => {
     async function getAllBoasts() {
       const res = await axios.get(
-        'http://ec2-3-34-95-255.ap-northeast-2.compute.amazonaws.com:8080/boasts'
+        `http://ec2-3-34-95-255.ap-northeast-2.compute.amazonaws.com:8080/boasts?page=${page}&size=${size}`
       );
       let data = res.data;
       setList(data);
-      console.log(list);
+      console.log(data);
     }
     try {
       getAllBoasts();
@@ -23,11 +21,6 @@ function Boast() {
       console.error(err);
     }
   }, []);
-
-  const navigate = useNavigate();
-  const handleOnClick = (id) => {
-    navigate(`/boasts/${id}`);
-  };
 
   const Container = styled.main`
     display: flex;
@@ -37,14 +30,10 @@ function Boast() {
   `;
 
   return (
-    <>
-      <Container>
-        <TitleHeader title={'자랑할래요'} />
-        <Link to="/boast/id" onClick={() => handleOnClick(list.memberId)}>
-          <Card likeButton={true} />
-        </Link>
-      </Container>
-    </>
+    <Container>
+      <TitleHeader title={'자랑할래요'} />
+      <Card likeButton={true} list={list} />
+    </Container>
   );
 }
